% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{busy_poll_job}
\alias{busy_poll_job}
\title{Poll job's status}
\usage{
busy_poll_job(zoltar_connection, job_url, verbose = TRUE)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{job_url}{URL of a valid job in zoltar_connection}

\item{verbose}{if TRUE, print messages on job status poll}
}
\description{
A convenience function that polls the passed Job's status waiting for either FAILED, TIMEOUT, or SUCCESS.
}
\examples{
\dontrun{
  busy_poll_job(conn, "http://example.com/api/job/2/")
}
}
