% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create a project}
\usage{
create_project(zoltar_connection, project_config)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_config}{A `list` containing a Zoltar project configuration. note that this list validated by the
server and not here. An example: cdc-project.json Full documentation at \url{https://docs.zoltardata.com/}.}
}
\value{
project_url of the newly-created project
}
\description{
Creates the project using the passed project configuration list. Fails if a project with the passed name already
  exists.
}
\examples{
\dontrun{
  new_project_url <- create_project(conn, jsonlite::read_json("cdc-project.json"))
}
}
