% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{csv_filename}
\alias{csv_filename}
\title{get a forecast's csv_filename}
\usage{
csv_filename(forecast, ...)
}
\arguments{
\item{forecast}{a \code{\link{Forecast}} object}
}
\value{
filename string
}
\description{
A \emph{getter} function that returns the csv_filename of the passed \code{\link{Forecast}}.
}
