% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{data}
\alias{data}
\title{get a forecast's data}
\usage{
data(forecast, is_json = TRUE, ...)
}
\arguments{
\item{is_json}{a boolean specifying whether the forecast is in JSON (the default) or CSV format}
}
\value{
forecast data associated with the passed \code{\link{Forecast}} object based on the requested format
}
\description{
get a forecast's data
}
