% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{data_frame_from_forecast_data}
\alias{data_frame_from_forecast_data}
\title{Converts forecast data from Zoltar's native `list` format to a `data.frame`}
\usage{
data_frame_from_forecast_data(forecast_data)
}
\arguments{
\item{forecast_data}{Forecast data as a `list` in the Zoltar standard format}
}
\value{
A `data.frame` from forecast_data in zoltar-specific format. The 'class' of each row is
  as described in https://docs.zoltardata.com/datamodel/ and is one of: 'bin', 'named', 'point',
  'sample', 'quantile'. The columns are: 'unit', 'target', 'class', 'value', 'cat', 'prob',
  'sample', 'quantile', 'family', 'param1', 'param2', 'param3'. Note that the frame is 'sparse':
  not every row uses all columns, and unused ones are empty (''). However, the first four
  columns are always non-empty, i.e., every prediction has them. NB: columns are all text (i.e.,
  data type information from forecast_data is lost)
}
\description{
Converts forecast data from Zoltar's native `list` format to a `data.frame`
}
\examples{
\dontrun{
  forecast_data <- jsonlite::read_json("docs-predictions.json")
  data_frame <- data_frame_from_forecast_data(forecast_data)
}
}
