% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{delete_model}
\alias{delete_model}
\title{Delete a model}
\usage{
delete_model(zoltar_connection, model_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{model_url}{URL of a model in zoltar_connection's models}
}
\value{
None
}
\description{
Deletes the model with the passed ID. This is permanent and cannot be undone.
}
\examples{
\dontrun{
  delete_model(conn, "http://www.zoltardata.com/api/model/1/")
}
}
