% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project}
\usage{
delete_project(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
None
}
\description{
Deletes the project with the passed URL. This is permanent and cannot be undone.
}
\examples{
\dontrun{
  delete_project(conn, "https://www.zoltardata.com/project/9/")
}
}
