% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{do_zoltar_query}
\alias{do_zoltar_query}
\title{A convenience function to construct and execute a Zoltar query for all quantile
forecasts submitted to a project within a specified number of days of a
given date.}
\usage{
do_zoltar_query(
  zoltar_connection,
  project_url,
  models = NULL,
  units = NULL,
  targets = NULL,
  timezeros = NULL,
  types = NULL,
  verbose = TRUE
)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}

\item{models}{Character vector of model abbreviations}

\item{units}{character vector of units to retrieve, e.g., c("01003", "US")}

\item{targets}{character vector of targets to retrieve, for example
c("1 wk ahead cum death", "2 wk ahead cum death")}

\item{timezeros}{character vector of timezeros to retrieve in YYYY_MM_DD_DATE_FORMAT, e.g., '2017-01-17'}

\item{types}{character vector of types as documented at at \url{https://docs.zoltardata.com/}}

\item{verbose}{if TRUE, print messages on job status poll}
}
\value{
A `data.frame` of Job's data. Full documentation at \url{https://docs.zoltardata.com/}.
}
\description{
A convenience function to construct and execute a Zoltar query for all quantile
forecasts submitted to a project within a specified number of days of a
given date.
}
\examples{
\dontrun{
  forecast_data <- do_zoltar_query(
    conn, "https://www.zoltardata.com/api/project/44/",
    c("CMU-TimeSeries", "UMass-MechBayes"), c("01003", "US"), c("1 wk ahead inc death"),
    c("2020-07-19", "2020-07-20"), c("quantile"))
}
}
