% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{download_forecast}
\alias{download_forecast}
\title{Gets a forecast's data}
\usage{
download_forecast(zoltar_connection, forecast_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{forecast_url}{URL of a forecast in zoltar_connection's forecasts}
}
\value{
Forecast data as a `list` in the Zoltar standard format - see \url{https://www.zoltardata.com/docs#forecasts}.
  meta information is ignored.
}
\description{
Gets a forecast's data
}
\examples{
\dontrun{
  forecast_data <- download_forecast(conn, "http://example.com/api/forecast/1/")
}
}
