% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{forecast_data}
\alias{forecast_data}
\title{get a forecast's data}
\usage{
forecast_data(zoltar_connection, forecast_id, is_json)
}
\arguments{
\item{zoltar_connection}{a ZoltarConnection object as returned by \code{\link{new_connection}}}

\item{forecast_id}{id of a forecast in zoltar_connection's forecasts}

\item{is_json}{a boolean specifying whether the forecast is in JSON or CSV format}
}
\value{
forecast data in the requested format
}
\description{
get a forecast's data
}
