% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc.R
\name{forecast_data_from_cdc_data_frame}
\alias{forecast_data_from_cdc_data_frame}
\title{\code{\link[=forecast_data_from_cdc_csv_file]{forecast_data_from_cdc_csv_file()}} helper}
\usage{
forecast_data_from_cdc_data_frame(season_start_year, cdc_data_frame)
}
\arguments{
\item{season_start_year}{as passed to \code{\link[=forecast_data_from_cdc_csv_file]{forecast_data_from_cdc_csv_file()}}}

\item{cdc_data_frame}{""}
}
\value{
same as \code{\link[=forecast_data_from_cdc_csv_file]{forecast_data_from_cdc_csv_file()}}
}
\description{
\code{\link[=forecast_data_from_cdc_csv_file]{forecast_data_from_cdc_csv_file()}} helper
}
