% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{forecast_for_pk}
\alias{forecast_for_pk}
\title{get the forecast for an id}
\usage{
forecast_for_pk(model, forecast_pk, ...)
}
\arguments{
\item{model}{a \code{Model} object}

\item{forecast_pk}{the id of a forecast on the host. for example, the id for the model with the uri
\emph{http://example.com/api/forecast/71/} is 71.}
}
\value{
a \code{\link{Forecast}} object
}
\description{
Utility function that returns the \code{\link{Forecast}} object with the passed primary key (AKA id). Used for
example in the case where you've called \code{\link{upload_forecast}} to get an \code{\link{UploadFileJob}} object,
from which you then want to obtain the newly-uploaded Forecast (see the demo app) - once the upload is successfully
complete.
}
