% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{forecasts}
\alias{forecasts}
\title{get a model's forecasts}
\usage{
forecasts(model, ...)
}
\arguments{
\item{model}{a \code{Model} object}
}
\value{
a list of \code{\link{Forecast}} objects in the passed model.
}
\description{
get a model's forecasts
}
