% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{id}
\alias{id}
\title{A \emph{getter} function that returns the id of the passed \code{\link{Project}}, \code{\link{Model}}, or}
\usage{
id(zoltar_resource, ...)
}
\arguments{
\item{zoltar_resource}{the object to get the id of}
}
\value{
none
}
\description{
A \emph{getter} function that returns the id of the passed \code{\link{Project}}, \code{\link{Model}}, or
}
