% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_data}
\alias{job_data}
\title{Gets a job's file's data}
\usage{
job_data(zoltar_connection, job_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{job_url}{URL of a valid job in zoltar_connection that has a data file associated with it}
}
\value{
A `data.frame` of Job's data. Full documentation at \url{https://docs.zoltardata.com/}.
}
\description{
Downloads the data associated with a job that has an associated file, such as a query's results.
}
\examples{
\dontrun{
  the_job_data <- job_data(conn, "http://example.com/api/job/2/")
}
}
