% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_info_forecast_url}
\alias{job_info_forecast_url}
\title{Get a new forecast upload's url}
\usage{
job_info_forecast_url(zoltar_connection, the_job_info)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{the_job_info}{a \code{list} object as returned by \code{\link[=job_info]{job_info()}}}
}
\value{
A URL of the new forecast
}
\description{
A helper function for jobs representing file uploads. Returns the URL of a newly-uploaded forecast
from job_info.
}
\examples{
\dontrun{
  new_forecast_url <- job_info_forecast_url(conn, "http://example.com/api/job/2/")
}
}
