% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{latest_forecasts}
\alias{latest_forecasts}
\title{Get a project's latest_forecasts}
\usage{
latest_forecasts(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A \code{data.frame} of all of the latest forecasts for the passed project. columns: forecast_id, source.
(Later we may generalize to allow passing specific columns to retrieve, such as 'forecast_model_id',
'time_zero_id', 'issued_at', 'created_at', 'source', and 'notes'.)
}
\description{
Get a project's latest_forecasts
}
\examples{
\dontrun{
  the_latest_forecasts <- latest_forecasts(conn, "https://www.zoltardata.com/api/project/9/")
}
}
