% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{models}
\alias{models}
\title{Returns a list of \code{\link{Model}} objects in the passed project.}
\usage{
models(project, ...)
}
\arguments{
\item{project}{a \code{\link{Project}} object}
}
\value{
a list of \code{\link{Model}} objects
}
\description{
Returns a list of \code{\link{Model}} objects in the passed project.
}
