% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{name}
\alias{name}
\title{A \emph{getter} function that returns the name of the passed \code{\link{Project}}.}
\usage{
name(project, ...)
}
\arguments{
\item{zoltar_resource}{the object to get the id of}
}
\value{
none
}
\description{
A \emph{getter} function that returns the name of the passed \code{\link{Project}}.
}
