% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{new_project}
\alias{new_project}
\title{make a new project (lower level function)}
\usage{
new_project(zoltar_connection, url)
}
\arguments{
\item{zoltar_connection}{a ZoltarConnection object as returned by \code{\link{new_connection}}}

\item{url}{host address of the Project, e.g., \emph{http://example.com/api/project/1/}.}
}
\value{
a Project object
}
\description{
Constructor that returns a new Project object for a particular project's url. Not normally called
directly - most users use the \code{\link{projects}} function.
}
