% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{project_info}
\alias{project_info}
\title{get a project's information}
\usage{
project_info(zoltar_connection, project_id)
}
\arguments{
\item{zoltar_connection}{a ZoltarConnection object as returned by \code{\link{new_connection}}}

\item{project_id}{id of a project in zoltar_connection's projects}
}
\value{
a list of project information for the passed project_id
}
\description{
get a project's information
}
