% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{projects}
\alias{projects}
\title{get a list of all projects}
\usage{
projects(zoltar_connection, ...)
}
\arguments{
\item{zoltar_connection}{a ZoltarConnection object as returned by \code{\link{new_connection}}}

\item{...}{ignored}
}
\value{
a list of Project objects
}
\description{
Returns a list of Project objects, which are the starting point for working with the API.
}
