% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{refresh}
\alias{refresh}
\title{update a resource}
\usage{
refresh(zoltar_resource, ...)
}
\arguments{
\item{zoltar_resource}{the object to refresh}

\item{...}{ignored}
}
\value{
none
}
\description{
Updates the internal (JSON) data associated with a Project, Model, Forecast,
or UploadFileJob object. Refreshing is required when you've changed something on the server that would
change what the API would return after the change. For example, creating a new Project on the site means the list
that \code{\link{projects}} returns would be stale and need updating. You'll also need to refresh after changing an
object's contents on the site, such as a model name or its forecasts. The API works this way, instead of
auto-refresh, because the latter would require querying the server every time.
}
