% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{scores}
\alias{scores}
\title{get a project's scores}
\usage{
scores(project, ...)
}
\value{
score data associated for all models in the passed \code{\link{Project}} object in CSV format
}
\description{
get a project's scores
}
