% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{status_as_str}
\alias{status_as_str}
\title{get an upload_file_job's int status as a string}
\usage{
status_as_str(upload_file_job, ...)
}
\arguments{
\item{upload_file_job}{an \code{\link{UploadFileJob}} object}
}
\value{
the passed \code{\link{UploadFileJob}}'s status as a human-readable string (rather than the internal int).
 the possible values are: "PENDING", "CLOUD_FILE_UPLOADED", "QUEUED", "CLOUD_FILE_DOWNLOADED", "SUCCESS", or
 "FAILED".
}
\description{
get an upload_file_job's int status as a string
}
