% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{target_info}
\alias{target_info}
\title{Get information about a target}
\usage{
target_info(zoltar_connection, target_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{target_url}{URL of a target in zoltar_connection's targets}
}
\value{
A \code{list} of target information for the passed target_url
}
\description{
Get information about a target
}
\examples{
\dontrun{
  the_target_info <- target_info(conn, "https://www.zoltardata.com/api/target/3/")
}
}
