% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{timezero_info}
\alias{timezero_info}
\title{Get information about a timezero}
\usage{
timezero_info(zoltar_connection, timezero_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{timezero_url}{URL of a timezero in zoltar_connection's timezeros}
}
\value{
A `list` of timezero information for the passed timezero_url
}
\description{
Get information about a timezero
}
\examples{
\dontrun{
  the_timezero_info <- timezero_info(conn, "https://www.zoltardata.com/timezero/3/")
}
}
