% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{truth}
\alias{truth}
\title{Get a project's truth}
\usage{
truth(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A `data.frame` of truth data for the passed project URL
}
\description{
Get a project's truth
}
\examples{
\dontrun{
  the_truth <- truth(conn, "https://www.zoltardata.com/project/9/")
}
}
