% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{truth_info}
\alias{truth_info}
\title{Get information about a project's truth}
\usage{
truth_info(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A \code{list} of project truth information for the passed project_url
}
\description{
Get information about a project's truth
}
\examples{
\dontrun{
  the_truth_info <- truth_info(conn, "https://www.zoltardata.com/api/project/9/")
}
}
