% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{unit_info}
\alias{unit_info}
\title{Get information about a unit}
\usage{
unit_info(zoltar_connection, unit_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{unit_url}{URL of a unit in zoltar_connection's zoltar_units}
}
\value{
A \code{list} of unit information for the passed unit_url
}
\description{
Get information about a unit
}
\examples{
\dontrun{
  the_unit_info <- unit_info(conn, "https://www.zoltardata.com/api/unit/3/")
}
}
