% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{upload_info}
\alias{upload_info}
\title{Get an upload's information}
\usage{
upload_info(zoltar_connection, upload_file_job_id)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{upload_file_job_id}{ID of a job in zoltar_connection that was uploaded via \code{\link{upload_forecast}}}
}
\value{
A `list` of upload information for the passed upload_file_job_id. it has these names:
  id, url, status, user, created_at, updated_at, failure_message, filename, input_json, output_json
}
\description{
Gets an upload's information that can be used to track the upload's progress. (Uploads are processed in a queue,
}
\examples{
\dontrun{
  the_upload_info <- upload_info(conn, 287L)
}
}
