% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-file-job.R
\name{upload_info_forecast_url}
\alias{upload_info_forecast_url}
\title{Get a new forecast upload's url}
\usage{
upload_info_forecast_url(zoltar_connection, the_upload_info)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{the_upload_info}{a `list` object as returned by \code{\link{upload_info}}}
}
\value{
A URL of the new forecast
}
\description{
A helper function that returns the URL of a newly-uploaded forecast from upload_info.
}
\examples{
\dontrun{
  new_forecast_url <- upload_info_forecast_url(conn, "http://example.com/api/uploadfilejob/2/")
}
}
