% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{z_authenticate}
\alias{z_authenticate}
\title{log in to a Zoltar host}
\usage{
z_authenticate(zoltar_connection, username, password, ...)
}
\arguments{
\item{zoltar_connection}{a \code{\link{ZoltarConnection}} object as returned by \code{\link{new_connection}}}

\item{username}{username for the account to use on the connection's host}

\item{password}{password ""}
}
\value{
none
}
\description{
Returns a new \code{\link{ZoltarConnection}} object, which is the starting point for working with the Zoltar API.
Once you have the connection you can call z_authenticate() on it, and call projects() to get a list of objects to
start working with.
}
