% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{zoltar_units}
\alias{zoltar_units}
\title{Get a project's zoltar_units}
\usage{
zoltar_units(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A `data.frame` of unit contents for the passed project
}
\description{
Get a project's zoltar_units
}
\examples{
\dontrun{
  the_units <- zoltar_units(conn, "https://www.zoltardata.com/project/9/")
}
}
