% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoltr.R
\docType{package}
\name{zoltr-package}
\alias{zoltr}
\alias{zoltr-package}
\title{zoltr: Interface to the 'Zoltar' Forecast Repository API}
\description{
'Zoltar' \url{https://www.zoltardata.com/} is a website that provides a repository of model forecast results in a standardized format and a central location. It supports storing, retrieving, comparing, and analyzing time series forecasts for prediction challenges of interest to the modeling community. This package provides functions for working with the 'Zoltar' API, including connecting and authenticating, getting meta information (projects, models, and forecasts, and truth), and uploading, downloading, and deleting forecast and truth data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reichlab/zoltr }
  \item \url{http://reichlab.io/zoltr/}
  \item Report bugs at \url{https://github.com/reichlab/zoltr/issues}
}

}
\author{
\strong{Maintainer}: Matthew Cornell \email{cornell@umass.edu}

Authors:
\itemize{
  \item Nicholas Reich \email{nick@schoolph.umass.edu} [copyright holder]
}

}
\keyword{internal}
