% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{data_frame_from_forecast_data}
\alias{data_frame_from_forecast_data}
\title{Converts forecast data from Zoltar's native `list` format to a `data.frame`}
\usage{
data_frame_from_forecast_data(forecast_data)
}
\arguments{
\item{forecast_data}{Forecast data as a `list` in the Zoltar standard format}
}
\value{
A `data.frame` from forecast_data in zoltar-specific format. The columns are:
  'unit', 'target', 'class', 'value', 'cat', 'prob', 'sample', 'quantile', 'family',
  'param1', 'param2', 'param3'. They are documented at
  https://docs.zoltardata.com/fileformats/#forecast-data-format-csv .
  NB: columns are all character (i.e., data type information from forecast_data is lost)
}
\description{
Converts forecast data from Zoltar's native `list` format to a `data.frame`
}
\examples{
\dontrun{
  forecast_data <- jsonlite::read_json("docs-predictions.json")
  data_frame <- data_frame_from_forecast_data(forecast_data)
}
}
