% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{delete_forecast}
\alias{delete_forecast}
\title{Delete a forecast}
\usage{
delete_forecast(zoltar_connection, forecast_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{forecast_url}{URL of a forecast in zoltar_connection's forecasts}
}
\value{
None
}
\description{
Deletes the forecast with the passed URL. This is permanent and cannot be undone.
}
\examples{
\dontrun{
  delete_forecast(conn, "http://example.com/api/forecast/1/")
}
}
