% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{forecast_data}
\alias{forecast_data}
\title{get a forecast's data}
\usage{
forecast_data(forecast, is_json = TRUE, ...)
}
\arguments{
\item{forecast}{a Forecast object}

\item{is_json}{a boolean specifying whether the forecast is in JSON (the default) or CSV format}

\item{...}{ignored}
}
\value{
forecast data associated with the passed Forecast object based on the requested format
}
\description{
get a forecast's data
}
