% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{forecast_data}
\alias{forecast_data}
\title{Gets a forecast's data}
\usage{
forecast_data(zoltar_connection, forecast_id, is_json)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{forecast_id}{ID of a forecast in zoltar_connection's forecasts}

\item{is_json}{A boolean specifying whether the forecast is in `list` or `data.frame` format}
}
\value{
Forecast data in the requested format - either a `list` or a `data.frame`
}
\description{
Gets a forecast's data
}
\examples{
\dontrun{
  forecast_data_json <- forecast_data(conn, 46L, is_json=TRUE)
  forecast_data_csv <- forecast_data(conn, 46L, is_json=FALSE)
}
}
