% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_info}
\alias{forecast_info}
\title{Gets a forecast's information}
\usage{
forecast_info(zoltar_connection, forecast_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{forecast_url}{URL of a forecast in zoltar_connection's forecasts}
}
\value{
A `list` of forecast information for the passed forecast_url
}
\description{
Gets a forecast's information
}
\examples{
\dontrun{
  the_forecast_info <- forecast_info(conn, "http://example.com/api/forecast/1/")
}
}
