% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecasts}
\alias{forecasts}
\title{Get a model's forecasts}
\usage{
forecasts(zoltar_connection, model_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{model_url}{URL of a model in zoltar_connection's models}
}
\value{
A `data.frame` of forecast information for the passed model
}
\description{
Get a model's forecasts
}
\examples{
\dontrun{
  the_forecasts <- forecasts(conn, "http://www.zoltardata.com/api/model/1/")
}
}
