% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_resource}
\alias{get_resource}
\title{Get JSON for a resource (URL). Authenticates if necessary}
\usage{
get_resource(zoltar_connection, url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{url}{A string of the resource's URL}
}
\value{
A `list` that contiains JSON information for the passed URL
}
\description{
Get JSON for a resource (URL). Authenticates if necessary
}
