% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{id}
\alias{id}
\title{get a resource's id}
\usage{
id(zoltar_resource, ...)
}
\arguments{
\item{zoltar_resource}{the object to get the id of}

\item{...}{ignored}
}
\value{
none
}
\description{
A \emph{getter} function that returns the id of the passed Project, Model, or
}
