% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_info}
\alias{model_info}
\title{Get information about a model}
\usage{
model_info(zoltar_connection, model_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{model_url}{URL of a model in zoltar_connection's models}
}
\value{
A \code{list} of model information for the passed model_url
}
\description{
Get information about a model
}
\examples{
\dontrun{
  the_model_info <- model_info(conn, "http://www.zoltardata.com/api/model/1/")
}
}
