% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{model_info}
\alias{model_info}
\title{Get information about a model}
\usage{
model_info(zoltar_connection, model_id)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{model_id}{ID of a model in zoltar_connection's models}
}
\value{
A `list` of model information for the passed model_id
}
\description{
Get information about a model
}
\examples{
\dontrun{
  the_model_info <- model_info(conn, 26L)
}
}
