% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{models}
\alias{models}
\title{get a project's models}
\usage{
models(project, ...)
}
\arguments{
\item{project}{a Project object}

\item{...}{ignored}
}
\value{
a list of Model objects
}
\description{
Returns a list of Model objects in the passed project.
}
