% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{models}
\alias{models}
\title{Get a project's models}
\usage{
models(zoltar_connection, project_id)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_id}{ID of a project in zoltar_connection's projects}
}
\value{
A `data.frame` of model contents for all models in the passed project
}
\description{
Get a project's models
}
\examples{
\dontrun{
  the_models <- models(conn, 4L)
}
}
