% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{models}
\alias{models}
\title{Get a project's models}
\usage{
models(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A \code{data.frame} of model contents for all models in the passed project
}
\description{
Get a project's models
}
\examples{
\dontrun{
  the_models <- models(conn, "https://www.zoltardata.com/api/project/9/")
}
}
