% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{new_connection}
\alias{new_connection}
\title{Get a connection to a Zoltar host}
\usage{
new_connection(host = "https://zoltardata.com")
}
\arguments{
\item{host}{The Zoltar site to connect to. Defaults to \url{https://zoltardata.com}}
}
\value{
A `ZoltarConnection` object
}
\description{
Returns a new connection object, which is the starting point for working with the Zoltar API. Once you have the
connection you can call \code{\link{zoltar_authenticate}} on it, and then call \code{\link{projects}} to get a list
of Project objects to start working with.
}
\examples{
\dontrun{
  conn <- new_connection()
}
}
