% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{new_model}
\alias{new_model}
\title{make a new model (lower level function)}
\usage{
new_model(zoltar_connection, url)
}
\arguments{
\item{zoltar_connection}{a ZoltarConnection object as returned by \code{\link{new_connection}}}

\item{url}{host address of the Model, e.g., \emph{http://example.com/api/model/1/}.}
}
\value{
a Model object
}
\description{
Constructor that returns a new Model object for a particular model's url. Not normally called
directly.
}
