% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_info}
\alias{project_info}
\title{Get information about a project}
\usage{
project_info(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A \code{list} of project information for the passed project_url
}
\description{
Get information about a project
}
\examples{
\dontrun{
  the_project_info <- project_info(conn, "https://www.zoltardata.com/api/project/9/")
}
}
