% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{projects}
\alias{projects}
\title{Returns a list of \code{\link{Project}} objects, which are the starting point for working with the API.}
\usage{
projects(zoltar_connection, ...)
}
\arguments{
\item{zoltar_connection}{a \code{ZoltarConnection} object as returned by \code{\link{new_connection}}}
}
\value{
a list of \code{\link{Project}} objects
}
\description{
Returns a list of \code{\link{Project}} objects, which are the starting point for working with the API.
}
