% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{projects}
\alias{projects}
\title{get a data.frame of all projects' metadata}
\usage{
projects(zoltar_connection)
}
\arguments{
\item{zoltar_connection}{a ZoltarConnection object as returned by \code{\link{new_connection}}}
}
\value{
a data.frame of project contents
}
\description{
Returns a data.frame of all projects' contents.
}
