% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{projects}
\alias{projects}
\title{Get information about all projects}
\usage{
projects(zoltar_connection)
}
\arguments{
\item{zoltar_connection}{A \code{ZoltarConnection} object as returned by \code{\link[=new_connection]{new_connection()}}}
}
\value{
A \code{data.frame} of all projects' contents
}
\description{
Get information about all projects
}
\examples{
\dontrun{
  the_projects <- projects(conn)
}
}
