% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{query_with_ids}
\alias{query_with_ids}
\title{Prepare a query}
\usage{
query_with_ids(zoltar_connection, project_url, query)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}

\item{query}{A `list` as documented in \code{\link{submit_query}}, but contains strings as
described above, not IDs}
}
\value{
a copy of query that has IDs substituted for strings
}
\description{
A convenience function that prepares a query for \code{\link{submit_query}} by replacing strings
with database IDs. Replaces these strings: "models": model_abbr -> ID. "units": unit_name -> ID.
"targets": target_name -> ID. "timezeros" timezero_date in YYYY_MM_DD_DATE_FORMAT-> ID.
}
\examples{
\dontrun{
  prepared_query <- query_with_ids(
    conn, "https://www.zoltardata.com/api/project/9/",
    list("models" = c("CMU-TimeSeries", "UMass-MechBayes"), "units" = c("US", "01"),
         "targets" = c("0 day ahead cum death", "1 wk ahead inc death"),
         "timezeros" = c("2020-07-19", "2020-07-20"), "types" = c("point", "quantile")))
}
}
