% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{scores}
\alias{scores}
\title{Get a project's scores}
\usage{
scores(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A `data.frame` of score data for all models in the passed project URL
}
\description{
Get a project's scores
}
\examples{
\dontrun{
  the_scores <- scores(conn, "https://www.zoltardata.com/project/9/")
}
}
