% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{scores}
\alias{scores}
\title{Get a project's scores}
\usage{
scores(zoltar_connection, project_id)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_id}{ID of a project in zoltar_connection's projects}
}
\value{
A `data.frame` of score data for all models in the passed project ID
}
\description{
Get a project's scores
}
\examples{
\dontrun{
  the_scores <- scores(conn, 4L)
}
}
