% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{targets}
\alias{targets}
\title{Get a project's targets}
\usage{
targets(zoltar_connection, project_url)
}
\arguments{
\item{zoltar_connection}{A `ZoltarConnection` object as returned by \code{\link{new_connection}}}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
A `data.frame` of target contents for the passed project
}
\description{
Get a project's targets
}
\examples{
\dontrun{
  the_targets <- targets(conn, "https://www.zoltardata.com/project/9/")
}
}
