% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{timezero_date}
\alias{timezero_date}
\title{get a forecast's timezero_date}
\usage{
timezero_date(forecast, ...)
}
\arguments{
\item{forecast}{a Forecast object}

\item{...}{ignored}
}
\value{
string formatted as YYYYMMDD
}
\description{
A \emph{getter} function that returns the timezero_date of the passed Forecast.
}
