% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{timezero_date}
\alias{timezero_date}
\title{A \emph{getter} function that returns the timezero_date of the passed \code{\link{Forecast}}.}
\usage{
timezero_date(forecast, ...)
}
\arguments{
\item{forecast}{a \code{\link{Forecast}} object}
}
\value{
string formatted as \code{YYYYMMDD}
}
\description{
A \emph{getter} function that returns the timezero_date of the passed \code{\link{Forecast}}.
}
